Move ConfigRedactor to org.apache.fluss.server.config packages
